+++ app/src/hooks/useSmartAccount.tsx
@@ -4,18 +4,14 @@ import { createSmartAccountClient } from "@alchemy/aa-core";
 import { baseSepolia, optimismSepolia, sepolia } from "@alchemy/aa-core";
 import { useMemo } from "react";
 import { http } from "viem";
-import { usePublicClient } from "wagmi";
 
 import { getRpcUrl } from "@/lib/alchemy";
 
-import { addresses } from "../../../contracts/shared/app/addresses";
 import { config } from "../../../contracts/shared/app/config";
 import { isChainId } from "../../../contracts/shared/app/types";
 
-export function useChain(chainId?: number) {
-  const chainPublicClient = usePublicClient({ chainId });
-
-  const chainSmartAccountClient = useMemo(() => {
+export function useSmartAccount(chainId?: number) {
+  const smartAccountClient = useMemo(() => {
     const _chainId = chainId?.toString();
     if (!isChainId(_chainId)) {
       console.log("not a chainId");
@@ -38,26 +34,7 @@ export function useChain(chainId?: number) {
     });
   }, [chainId]);
 
-  const chainConfig = useMemo(() => {
-    const _chainId = chainId?.toString();
-    if (!isChainId(_chainId)) {
-      return;
-    }
-    return config[_chainId];
-  }, [chainId]);
-
-  const chainAddresses = useMemo(() => {
-    const _chainId = chainId?.toString();
-    if (!isChainId(_chainId)) {
-      return;
-    }
-    return addresses[_chainId];
-  }, [chainId]);
-
   return {
-    chainPublicClient,
-    chainSmartAccountClient,
-    chainConfig,
-    chainAddresses,
+    smartAccountClient,
   };
 }
